refactor NumberSignal not to backed by EventQueue
